全てのIntegerからLongにする修正

+++ app/src/main/java/jp/wako/demo/springbootmvc/domain/issues/IIssueRepository.java
@@ -3,7 +3,7 @@ package jp.wako.demo.springbootmvc.domain.issues;
 import java.util.Optional;
 
 public interface IIssueRepository {
-    Optional<Issue> findById(final Integer id);
-    Integer save(final Issue issue);
+    Optional<Issue> findById(final Long id);
+    Long save(final Issue issue);
     void delete(final Issue issue);
 }
