+++ app/src/main/java/jp/wako/demo/springbootmvc/domain/shared/Entity.java
@@ -10,8 +10,8 @@ import lombok.Getter;
 @Getter
 @EqualsAndHashCode(of = {"id"})
 public class Entity {
-    private final Integer id;
+    private final Long id;
     private final LocalDateTime createdAt;
     protected LocalDateTime updatedAt;
-    private final Integer version;
+    private final Long version;
 }
