+++ app/src/main/java/jp/wako/demo/springbootmvc/infra/issues/IssueSearchCondition.java
@@ -5,7 +5,7 @@ import lombok.Data;
 
 @Data
 public final class IssueSearchCondition {
-    private final Integer projectId;
+    private final Long projectId;
     private final String title;
     private final Sort sort;
 }
