+++ app/src/main/java/jp/wako/demo/springbootmvc/infra/shared/dao/ImmutableEntity.java
@@ -16,5 +16,5 @@ public abstract class ImmutableEntity {
     private final LocalDateTime updatedAt;
 
     @Version
-    private final Integer version;
+    private final Long version;
 }
