+++ app/src/main/java/jp/wako/demo/springbootmvc/usecase/issues/create/IssueCreateRequest.java
@@ -4,7 +4,7 @@ import lombok.Data;
 
 @Data
 public final class IssueCreateRequest {
-    private final Integer projectId;
+    private final Long projectId;
     private final String title;
     private final String description;
 }
