+++ app/src/main/java/jp/wako/demo/springbootmvc/usecase/issues/search/IssueDto.java
@@ -4,8 +4,8 @@ import lombok.Data;
 
 @Data
 public final class IssueDto {
-    private final Integer id;
-    private final Integer projectId;
+    private final Long id;
+    private final Long projectId;
     private final String title;
     private final String description;
 }
