+++ app/src/test/java/jp/wako/demo/springbootmvc/usecase/issues/TestIssueFactory.java
@@ -8,49 +8,49 @@ public class TestIssueFactory {
 
     public static Issue create() {
         var issue = Issue.reconstruct(
-            1000,
-            1000,
+            1000L,
+            1000L,
             "Issue1",
             "This is a test issue.",
             LocalDateTime.now().withNano(0),
             LocalDateTime.now().withNano(0),
-            1);
+            1L);
         return issue;
     }
 
-    public static Issue create(final Integer id) {
+    public static Issue create(final Long id) {
         var issue = Issue.reconstruct(
             id,
-            1000,
+            1000L,
             "Issue1",
             "This is a test issue.",
             LocalDateTime.now().withNano(0),
             LocalDateTime.now().withNano(0),
-            1);
+            1L);
         return issue;
     }
 
     public static Issue create(final LocalDateTime createdAt, final LocalDateTime updatedAt) {
         var issue = Issue.reconstruct(
-            1000,
-            1000,
+            1000L,
+            1000L,
             "Issue1",
             "This is a test issue.",
             createdAt,
             updatedAt,
-            1);
+            1L);
         return issue;
     }
 
-    public static Issue create(final Integer id, final LocalDateTime createdAt, final LocalDateTime updatedAt) {
+    public static Issue create(final Long id, final LocalDateTime createdAt, final LocalDateTime updatedAt) {
         var issue = Issue.reconstruct(
             id,
-            1000,
+            1000L,
             "Issue1",
             "This is a test issue.",
             createdAt,
             updatedAt,
-            1);
+            1L);
         return issue;
     }
